+++ include/program/stratum.h
@@ -39,7 +39,7 @@ public:
                                bool is_negated,
                                std::vector<formula::Formula *> formula_vector);
 
-    bool evaluate(unsigned long long int time_point, unsigned long long int tuple_counter,
+    bool evaluate(uint64_t time_point, uint64_t tuple_counter,
               std::unordered_map<std::string, std::vector<formula::Formula *>>
               new_facts);
