+++ src/rule/rule.cpp
@@ -53,8 +53,8 @@ std::string Rule::get_head_predicate() const {
 }
 
 bool Rule::evaluate(
-        unsigned long long int current_time,
-        unsigned long long int current_tuple_counter) {}
+        uint64_t current_time,
+        uint64_t current_tuple_counter) {}
 
 //const std::unordered_map<std::string, std::vector<formula::Formula *>>
 //Rule::get_body_variable_map() const {
@@ -72,8 +72,8 @@ Rule::get_body_negated_predicate_map() const {
 }
 
 void Rule::expire_outdated_groundings(
-        unsigned long long int current_time,
-        unsigned long long int current_tuple_counter) {
+        uint64_t current_time,
+        uint64_t current_tuple_counter) {
     head.expire_outdated_groundings(current_time, current_tuple_counter);
     body.expire_outdated_groundings(current_time, current_tuple_counter);
